+++ CHANGELOG.md
@@ -3,14 +3,41 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v1.0.0 - [date]
+## v1.0.0 - 2023-05-22
 
-Initial release of nf-core/rnasplice, created with the [nf-core](https://nf-co.re/) template.
+First release of nf-core/rnasplice, created with the [nf-core](https://nf-co.re/) template.
 
 ### `Added`
 
-### `Fixed`
+Implemented pipeline:
 
-### `Dependencies`
+- Merge re-sequenced FastQ files (cat)
+- Read QC (FastQC)
+- Adapter and quality trimming (TrimGalore)
+- Alignment with STAR:
+  - STAR -> Salmon
+  - STAR -> featureCounts
+  - STAR -> HTSeq (DEXSeq count)
+- Sort and index alignments (SAMtools)
+- Create bigWig coverage files (BEDTools, bedGraphToBigWig)
+- Pseudo-alignment and quantification (Salmon; optional)
+- Summarize QC (MultiQC)
+- Differential Exon Usage (DEU):
+  - HTSeq -> DEXSeq
+  - featureCounts -> edgeR
+  - Quantification with featureCounts or HTSeq
+- Differential exon usage with DEXSeq or edgeR
+  - Differential Transcript Usage (DTU):
+  - Salmon -> DRIMSeq -> DEXSeq
+  - Filtering with DRIMSeq
+- Differential transcript usage with DEXSeq
+- Event-based splicing analysis:
+  - STAR -> rMATS
+  - Salmon -> SUPPA2
 
-### `Deprecated`
+Updated pipeline:
+
+- Visualization of differential results with edgeR, DEXSeq, and MISO
+- Contrasts specified using contrastsheet.csv
+- Allow users to specify input data type and start point (e.g., fastq, genome_bam, transcript_bam, salmon_results)
+- Pipeline schematic updated